+++ CHANGELOG.md
@@ -34,6 +34,9 @@
 
 ### Changes
 - For English, changed `Expert`'s difficulty level rename from `Bone Me Plenty` to `Turn Me Undead`. For obvious reasons.
+### Configuration
+- Added the following new options:
+  **Guns:** `EnableMinigunDynamicFirerate`, `EnableStarCannonDynamicFirerate`.
 ### Localization
 - Chinese (`59.1%` -> `100%`) - PR [#232](https://github.com/Mirsario/TerrariaOverhaul/pull/232) by [**xiaoyvyv**](https://github.com/xiaoyvyv).
 - Fixed Overhaul's main menu buttons sometimes being incorrectly shown in English when a different language is selected.